Create LICENSE
As most of the project is based and mixed with OSMBasicParser which is under GPL3, it uses the same licence.
If anybody would like to do a commercial usage of it, I strongly believe it better to rewrite the useful part. Feel free to contact me.
The GPX creation is based on trying to get a contiguous path on all selected segment. It is a "smart" solution to a stupid issue (as the client could really easily support non-contiguous segments), so shouldn't be of any use in the future.
The decimation algorithm can be used for non-contiguous segments though.